requirements.txt
streamlit
requests
torch
rdkit-pypi
py3Dmol
numpy
pandas
pillow
